feat: Category CRUD completed
Category CRUD is now completed successfully but whle deletion we still need to check weatehr that category have products or not. If have, so first need to show a modal with a msg like 'Are you sure that you want to delete that category?'. Something like that but in order to implement that we need to assign our category to our product first.